菜单增加隐藏/显示、增加通过分类parentId查询分类

+++ src/main/java/com/xslgy/common/service/impl/CmsCategoryServiceImpl.java
@@ -4,9 +4,11 @@ import com.xslgy.common.entity.CmsCategory;
 import com.xslgy.common.repository.CmsCategoryRepository;
 import com.xslgy.common.service.CmsCategoryService;
 import com.xslgy.common.utils.PageUtils;
+import com.xslgy.core.exception.XSLGYException;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.domain.Specification;
 import org.springframework.stereotype.Service;
+import org.springframework.util.CollectionUtils;
 import org.springframework.util.StringUtils;
 
 import javax.annotation.Resource;
@@ -39,11 +41,24 @@ public class CmsCategoryServiceImpl implements CmsCategoryService {
 
     @Override
     public CmsCategory save(CmsCategory cmsCategory) {
+        if (cmsCategory.getParentId() == null) {
+            // parentId为空的时候，默认为顶级菜单
+            cmsCategory.setParentId(0L);
+        } else {
+            CmsCategory category = getById(cmsCategory.getParentId());
+            if (category == null) {
+                throw new XSLGYException("上级菜单不存在");
+            }
+        }
         return cmsCategoryRepository.save(cmsCategory);
     }
 
     @Override
     public void deleteById(Long id) {
+        List<CmsCategory> cmsCategories = getByParentId(id);
+        if (!CollectionUtils.isEmpty(cmsCategories)) {
+            throw new XSLGYException("该分类下还有子分类，请先删除子分类");
+        }
         cmsCategoryRepository.deleteById(id);
     }
 
@@ -60,5 +75,12 @@ public class CmsCategoryServiceImpl implements CmsCategoryService {
     @Override
     public List<CmsCategory> getByParentId(Long parentId) {
         return cmsCategoryRepository.findByParentIdOOrderByOrerNoDesc(parentId);
+    }
+
+    @Override
+    public CmsCategory hideOrShow(Long id, Integer isShow) {
+        CmsCategory category = getById(id);
+        category.setShow(isShow);
+        return save(category);
     }
 }
